Remove ReferenceTestRunner, RegressionTestRunner and "matchers"
This removes the "run all files of some kind in some folder" feature,
but it's not clear that was being used; and isn't parallel to
run-fast-tests.

At best, ScriptGenerator generated a set of bash scripts which would
perform the same task as done by GenericTestCase, leaving the 'analysis'
of whether the testcases passed/failed to someone else.
This isn't a bad idea.

At the same time, the code is untested, and all seems to be
duplicating/repeating the effort of each other part.
For now it's better that it's not in the repo.